refactor:useAuth 훅 이용해서 userId 가져오기
