+++ src/pages/user/UserInfo.tsx
@@ -3,6 +3,7 @@ import ConfirmModal from '@/components/modal/ConfirmModal';
 import BasicInfo from '@/components/mypage/UserInfo/BasicInfo';
 import ExtraInfo from '@/components/mypage/UserInfo/ExtraInfo';
 import PasswordInfo from '@/components/mypage/UserInfo/PasswordInfo';
+import useAuth from '@/hooks/useAuth';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { type FC, useState } from 'react';
@@ -10,7 +11,7 @@ import './userInfo.scss';
 
 const UserInfo: FC = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const userId = localStorage.getItem('userId');
+  const { userId, logout } = useAuth();
   const oauth = localStorage.getItem('oauth');
 
   if (!userId) {
@@ -28,8 +29,7 @@ const UserInfo: FC = () => {
   const { mutate: deleteMember } = useMutation(deleteUser, {
     onSuccess: () => {
       alert('회원탈퇴 완료');
-      localStorage.clear();
-      window.location.href = '/';
+      logout();
     },
     onError: (err) => {
       if (axios.isAxiosError(err)) {
